chore!: refactored types into a separate folder
