--- src/resources/data/forms.ts
@@ -1,8 +1,0 @@
-export const forms: Record<string, string> = {
-  arewethereyet: "https://forms.gle/TZREGetDcMRtZY7G7",
-  aurafarming: "https://www.instagram.com/reel/DMg0SvhARRN/",
-  "councilor-applications":
-    "https://www.youtube.com/watch?v=dQw4w9WgXcQ&list=RDdQw4w9WgXcQ&start_radio=1",
-  "awty-handbook":
-    "https://docs.google.com/document/d/1UotSzRsbD6umpQfN0rHvS7MDXIfaHf7s1IlC7veRaVs/edit?usp=sharing",
-};
