+++ src/types/hackathon/sponsors.ts
@@ -1,0 +1,18 @@
+export type Tier =
+  | "platinum"
+  | "gold"
+  | "silver"
+  | "bronze"
+  | "inkind"
+  | "schoolSponsor";
+
+export type Sponsor = {
+  name: string;
+  logo: string;
+};
+
+export type YearSponsors = Partial<Record<Tier, Sponsor[]>>;
+
+export type HackathonYearSponsorsProps = {
+  year: string | number;
+};
