+++ src/types/hackathon/winner.ts
@@ -1,0 +1,6 @@
+export type HackathonWinner = {
+  prizeIndex: number;
+  projectName: string;
+  teamMembers: string[];
+  projectUrl?: string;
+};
